Temporary fix
The problem was at least to be found in the use of the `RefObjectStore` for
the `JobsList` entries for each `AuthInfo` in the `JobManager`. As soon
as no update requests are active the `JobsList` will be popped from the
store and with it the timestamp of the last scheduler update. In the
next request a new `JobsList` instance will then be generated whose next
poll delay will be zero, because it is a new instance.